Implemented map view

+++ OnTheMap/Models/Responses/StudentLocationResponse.swift
@@ -11,8 +11,8 @@ struct Result : Codable {
     
     let firstName: String
     let lastName: String
-    let longitude: Float
-    let latitude: Float
+    let longitude: Double
+    let latitude: Double
     let mapString: String
     let mediaURL: String
     let uniqueKey: String
